Some more explanations and reorganization of files
